refactor: rewrited fn addToCart; changed saving in ssesion memory
